pyproject.toml
[project]
name = "scfm-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "anndata>=0.12.6",
    "bionemo>=0.3.1.post2",
    "bionemo-scdl>=0.1.3",
    "dill>=0.4.0",
    "hickle>=5.0.3",
    "hydra-core>=1.3.2",
    "leidenalg>=0.11.0",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pubchempy>=1.0.5",
    "pytorch-lightning>=2.6.0",
    "rdkit>=2025.9.3",
    "scib-metrics>=0.5.7",
    "torch>=2.7.0",
    "torch-geometric>=2.7.0",
    "transformers>=4.57.3",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.3.0",
    "pre-commit>=4.4.0",
    "pytest>=9.0.1",
    "ruff>=0.14.5",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
# Enable Pyflakes (F), pycodestyle (E, W), isort (I), and bugbear (B)
select = ["F", "E", "W", "I", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = "src"

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
